requirements.txt
anyio==4.9.0
certifi==2025.1.31
h11==0.14.0
httpcore==1.0.7
httpx==0.25.2
idna==3.10
psycopg2-binary==2.9.10
python-dotenv==1.0.0
python-telegram-bot==20.7
PyYAML==6.0.1
sniffio==1.3.1
